moves funcs to convert to json to formats dir
