+++ app/formats/json/LabelFormat.scala
@@ -1,7 +1,8 @@
 package formats.json
 
+import controllers.helper.GoogleMapsHelper
+import models.label.LabelTable.{LabelMetadata, LabelMetadataUserDash, LabelValidationMetadata}
 import java.sql.Timestamp
-
 import models.label._
 import play.api.libs.json._
 import play.api.libs.functional.syntax._
@@ -52,4 +53,102 @@ object LabelFormat {
       (__ \ "photographer_heading").write[Float] and
       (__ \ "photographer_pitch").write[Float]
   )(unlift(LabelTable.LabelCVMetadata.unapply _))
+
+  def validationLabelMetadataToJson(labelMetadata: LabelValidationMetadata): JsObject = {
+    Json.obj(
+      "label_id" -> labelMetadata.labelId,
+      "label_type" -> labelMetadata.labelType,
+      "gsv_panorama_id" -> labelMetadata.gsvPanoramaId,
+      "image_date" -> labelMetadata.imageDate,
+      "label_timestamp" -> labelMetadata.timestamp,
+      "heading" -> labelMetadata.heading,
+      "pitch" -> labelMetadata.pitch,
+      "zoom" -> labelMetadata.zoom,
+      "canvas_x" -> labelMetadata.canvasX,
+      "canvas_y" -> labelMetadata.canvasY,
+      "canvas_width" -> labelMetadata.canvasWidth,
+      "canvas_height" -> labelMetadata.canvasHeight,
+      "severity" -> labelMetadata.severity,
+      "temporary" -> labelMetadata.temporary,
+      "description" -> labelMetadata.description,
+      "user_validation" -> labelMetadata.userValidation.map(LabelValidationTable.validationOptions.get),
+      "tags" -> labelMetadata.tags
+    )
+  }
+
+  def labelMetadataWithValidationToJsonAdmin(labelMetadata: LabelMetadata): JsObject = {
+    Json.obj(
+      "label_id" -> labelMetadata.labelId,
+      "gsv_panorama_id" -> labelMetadata.gsvPanoramaId,
+      "tutorial" -> labelMetadata.tutorial,
+      "image_date" -> labelMetadata.imageDate,
+      "heading" -> labelMetadata.heading,
+      "pitch" -> labelMetadata.pitch,
+      "zoom" -> labelMetadata.zoom,
+      "canvas_x" -> labelMetadata.canvasXY._1,
+      "canvas_y" -> labelMetadata.canvasXY._2,
+      "canvas_width" -> labelMetadata.canvasWidth,
+      "canvas_height" -> labelMetadata.canvasHeight,
+      "audit_task_id" -> labelMetadata.auditTaskId,
+      "user_id" -> labelMetadata.userId,
+      "username" -> labelMetadata.username,
+      "timestamp" -> labelMetadata.timestamp,
+      "label_type_key" -> labelMetadata.labelTypeKey,
+      "label_type_value" -> labelMetadata.labelTypeValue,
+      "severity" -> labelMetadata.severity,
+      "temporary" -> labelMetadata.temporary,
+      "description" -> labelMetadata.description,
+      "user_validation" -> labelMetadata.userValidation.map(LabelValidationTable.validationOptions.get),
+      "num_agree" -> labelMetadata.validations("agree"),
+      "num_disagree" -> labelMetadata.validations("disagree"),
+      "num_notsure" -> labelMetadata.validations("notsure"),
+      "tags" -> labelMetadata.tags
+    )
+  }
+
+  // Has the label metadata excluding username, user_id, and audit_task_id.
+  def labelMetadataWithValidationToJson(labelMetadata: LabelMetadata): JsObject = {
+    Json.obj(
+      "label_id" -> labelMetadata.labelId,
+      "gsv_panorama_id" -> labelMetadata.gsvPanoramaId,
+      "tutorial" -> labelMetadata.tutorial,
+      "image_date" -> labelMetadata.imageDate,
+      "heading" -> labelMetadata.heading,
+      "pitch" -> labelMetadata.pitch,
+      "zoom" -> labelMetadata.zoom,
+      "canvas_x" -> labelMetadata.canvasXY._1,
+      "canvas_y" -> labelMetadata.canvasXY._2,
+      "canvas_width" -> labelMetadata.canvasWidth,
+      "canvas_height" -> labelMetadata.canvasHeight,
+      "timestamp" -> labelMetadata.timestamp,
+      "label_type_key" -> labelMetadata.labelTypeKey,
+      "label_type_value" -> labelMetadata.labelTypeValue,
+      "severity" -> labelMetadata.severity,
+      "temporary" -> labelMetadata.temporary,
+      "description" -> labelMetadata.description,
+      "user_validation" -> labelMetadata.userValidation.map(LabelValidationTable.validationOptions.get),
+      "num_agree" -> labelMetadata.validations("agree"),
+      "num_disagree" -> labelMetadata.validations("disagree"),
+      "num_notsure" -> labelMetadata.validations("notsure"),
+      "tags" -> labelMetadata.tags
+    )
+  }
+
+  def labelMetadataUserDashToJson(label: LabelMetadataUserDash): JsObject = {
+    Json.obj(
+      "label_id" -> label.labelId,
+      "gsv_panorama_id" -> label.gsvPanoramaId,
+      "heading" -> label.heading,
+      "pitch" -> label.pitch,
+      "zoom" -> label.zoom,
+      "canvas_x" -> label.canvasX,
+      "canvas_y" -> label.canvasY,
+      "canvas_width" -> label.canvasWidth,
+      "canvas_height" -> label.canvasHeight,
+      "label_type" -> label.labelType,
+      "time_validated" -> label.timeValidated,
+      "validator_comment" -> label.validatorComment,
+      "image_url" -> GoogleMapsHelper.getImageUrl(label.gsvPanoramaId, label.canvasWidth, label.canvasHeight, label.heading, label.pitch, label.zoom)
+    )
+  }
 }
